Estudio_Mariana_Freitas_Beauty/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
firebase-admin==6.2.0
twilio==8.10.0
python-dotenv==1.0.0
pytz==2023.3
Flask-WTF==1.1.1
email-validator==2.0.0.post2
